requirements.txt
streamlit==1.32.0
faiss-cpu==1.7.4
python-docx==1.0.1
numpy==1.24.4
langchain-community==0.0.16
langchain==0.1.4
langchain-openai==0.0.5
PyPDF2==3.0.1
python-dotenv==1.0.0
nltk==3.8.1
rouge-score==0.1.2
pytesseract==0.3.10
Pillow==10.0.0
google-api-python-client==2.107.0
google-generativeai==0.3.2
httpx==0.25.0
sentence-transformers==2.2.2